mailer/requirements.txt
pika
python-dotenv
jinja2
pydantic[email]
httpx